backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
google-generativeai>=0.3.0
